# openapi/v1/openapi.yaml
openapi: 3.1.0
info:
  title: ScamNemesis API
  description: |
    REST API for the ScamNemesis fraud reporting and tracking platform.

    ## Authentication
    The API supports three authentication methods:
    - **Bearer Token**: JWT access token in Authorization header
    - **HttpOnly Cookie**: Automatic for browser clients
    - **API Key**: X-API-Key header for programmatic access

    ## Rate Limiting
    - Standard users: 100 requests/hour per endpoint
    - Admin users: 1000 requests/hour (10x multiplier)
    - Auth endpoints: 20 requests/15min (brute force protection)

    ## Error Handling
    All errors follow RFC 7807 Problem Details format with additional fields.
  version: 1.0.0
  contact:
    name: ScamNemesis Support
    url: https://scamnemesis.com/contact
    email: support@scamnemesis.com
  license:
    name: Proprietary
    url: https://scamnemesis.com/terms
servers:
  - url: https://scamnemesis.com/api/v1
    description: Production server
  - url: https://staging.scamnemesis.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - BearerAuth: []
  - CookieAuth: []
  - ApiKeyAuth: []
paths:
...
  /admin/reports/{id}:
    get:
      tags: [Admin Reports]
      summary: Get report details (admin)
      operationId: adminGetReport
      security:
        - BearerAuth: [admin:read]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Full report details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminReportDetail'
        '404':
          description: Report not found
...
  /admin/reports/{id}/reject:
    post:
      tags: [Admin Reports]
      summary: Reject report
      operationId: rejectReport
      security:
        - BearerAuth: [admin:write]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 1000
              required: [reason]
      responses:
        '200':
          description: Report rejected
        '404':
          description: Report not found
...
  /admin/comments/{id}/approve:
    post:
      tags: [Admin Comments]
      summary: Approve comment
      operationId: approveComment
      security:
        - BearerAuth: [admin:write]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comment approved
...
  /admin/duplicates:
    get:
      tags: [Admin Reports]
      summary: List duplicate clusters
      operationId: listDuplicates
      security:
        - BearerAuth: [admin:read]
      responses:
        '200':
          description: Duplicate clusters
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: HttpOnly cookie authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
    WebhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: Webhook secret for external services
...
  schemas:
...
    AdminReportSummary:
      type: object
      properties:
        id:
          type: string
        publicId:
          type: string
        caseNumber:
          type: string
        status:
          type: string
        severity:
          type: string
          nullable: true
        fraudType:
          type: string
        summary:
          type: string
        financialLoss:
          type: object
          nullable: true
          properties:
            amount:
              type: number
            currency:
              type: string
        country:
          type: string
        reporter:
          type: object
          nullable: true
          properties:
            id:
              type: string
            email:
              type: string
            displayName:
              type: string
        moderatedBy:
          type: object
          nullable: true
          properties:
            id:
              type: string
            displayName:
              type: string
        counts:
          type: object
          properties:
            perpetrators:
              type: integer
            evidence:
              type: integer
            comments:
              type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
          nullable: true
        moderatedAt:
          type: string
          format: date-time
          nullable: true
    AdminReportDetail:
      allOf:
        - $ref: '#/components/schemas/AdminReportSummary'
        - type: object
          properties:
            description:
              type: string
            perpetrators:
              type: array
              items:
                type: object
            digitalFootprint:
              type: object
            financialInfo:
              type: object
            cryptoInfo:
              type: object
            companyInfo:
              type: object
            vehicleInfo:
              type: object
            evidence:
              type: array
              items:
                type: object
            trackingToken:
              type: string
            reporterEmail:
              type: string
            reporterName:
              type: string
            reporterPhone:
              type: string
...